New UI v3.0

+++ src/App.tsx
@@ -7,7 +7,6 @@ import { Outlet } from "react-router-dom";
 import { useAppContext } from "./Services/AppContext";
 import "./App.scss";
 import { useMsgAppContext } from "./Services/MessagesContextAndInterfaces/MessagesContext";
-// import ErrorBoundary from "./Services/ErrorBoundary";
 
 function App() {
   const { dispatch, state } = useAppContext();
@@ -43,10 +42,6 @@ function App() {
   //   localStorage.setItem(`yashAppData`, JSON.stringify(state));
   // }, [state]);
 
-  // useEffect(() => {
-  //   Cookies.set("yashAppData", JSON.stringify(state), { expires: 1 / 24 }); // Expires in 7 days
-  // }, [state]);
-
   useEffect(() => {
     // Check for the cookie ID
     const cookieId = Cookies.get("appId");
@@ -54,35 +49,49 @@ function App() {
 
     if (cookieId) {
       // Update localStorage if cookie ID is present
-      if (localStorageData) {
-        const parsedData = JSON.parse(localStorageData);
-        if (parsedData) {
+      try {
+        const parsedData = localStorageData
+          ? JSON.parse(localStorageData)
+          : null;
+
+        // Only update localStorage if the state has changed
+        if (
+          !parsedData ||
+          JSON.stringify(parsedData) !== JSON.stringify(state)
+        ) {
           localStorage.setItem("yashAppData", JSON.stringify(state));
         }
-      } else {
-        localStorage.setItem("yashAppData", JSON.stringify(state));
+      } catch (error) {
+        console.error("Error parsing localStorage data:", error);
+        localStorage.removeItem("yashAppData"); // Optionally clear corrupt data
       }
     } else {
       // Clear localStorage if cookie ID is not present
       localStorage.clear();
-      // Generate a new unique ID and set it in the cookie
+
+      // Generate a new unique ID and set it in the cookie (expires in 2 days)
       const newId = Date.now().toString();
-      Cookies.set("appId", newId, { expires: 2 / 24 }); // Set cookie for 7 days
+      Cookies.set("appId", newId, { expires: 2 });
+
+      // Set localStorage with the current state
       localStorage.setItem("yashAppData", JSON.stringify(state));
     }
   }, [state]);
 
   useEffect(() => {
-    sessionStorage.setItem("yashAppMsgData", JSON.stringify(messageState));
+    // Compare messageState with sessionStorage to avoid unnecessary updates
+    const existingMsgData = sessionStorage.getItem("yashAppMsgData");
+
+    if (existingMsgData !== JSON.stringify(messageState)) {
+      sessionStorage.setItem("yashAppMsgData", JSON.stringify(messageState));
+    }
   }, [messageState]);
 
   return (
-    // <ErrorBoundary>
     <div className="w-screen h-[100dvh]">
       <Toast ref={myToast} />
       <Outlet />
     </div>
-    // </ErrorBoundary>
   );
 }
